% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digit256.R
\name{.digit256}
\alias{.digit256}
\title{Extract the n-th digit of a duocentehexaquinquagesimal number}
\usage{
.digit256(x, d)
}
\arguments{
\item{x}{\code{integer(n)}}

\item{d}{\code{integer(1)} One of 0:3.  Other integers silently return \code{x}.}
}
\value{
For \code{b = 256}
if \deqn{x = a_0 + a_1b + a_2b^2 + a_3b^3}
then \code{.digit(x, d) = a_d}
}
\description{
Extract the n-th digit of a duocentehexaquinquagesimal number
}
