% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/street-types.R
\name{.permitted_street_type_ord}
\alias{.permitted_street_type_ord}
\title{Street types allowed.}
\usage{
.permitted_street_type_ord()
}
\value{
A character vector, the permitted street codes. In order of
(approximate) occurrence; more common street types appear in the head
of the vector.
}
\description{
Street types allowed.
}
