% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_address.R
\name{encode_address}
\alias{encode_address}
\title{Encode standard address}
\usage{
encode_address(StandardAddress)
}
\arguments{
\item{StandardAddress}{A \code{data.table} returned by
\code{standardize_address}.}
}
\description{
See \code{standardize_address}
}
