% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_flatNumberFirstLast.R
\name{extract_flatNumberFirstLast}
\alias{extract_flatNumberFirstLast}
\title{Extract the flat number, number first/last from an address}
\usage{
extract_flatNumberFirstLast(address)
}
\arguments{
\item{address}{A character vector from which the numbers are to be extracted.}
}
\value{
A \code{data.table} of three components: the flat number,
the number first, and number last.
}
\description{
Extract the flat number, number first/last from an address
}
