% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_XXX.R
\name{get_the_XXX}
\alias{get_the_XXX}
\title{Elements of addresses with THE}
\usage{
get_the_XXX(o = TRUE)
}
\description{
Some addresses contain 'THE' which is a string often at the
beginning of some street names.
}
