% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HashStreetName.R
\name{HashStreetName}
\alias{HashStreetName}
\title{Hash a street name quickly and accurately}
\usage{
HashStreetName(x)
}
\arguments{
\item{x}{A character vector of uppercase street names (without the street type).}
}
\description{
Hash a street name quickly and accurately
}
\examples{
HashStreetName("FLINDERS")

}
