% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthyAddress-package.R
\docType{package}
\name{healthyAddress-package}
\alias{healthyAddress}
\alias{healthyAddress-package}
\title{Package for address standardization}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Hugh Parsonage \email{hugh.parsonage@gmail.com.com}

}
\keyword{internal}
