% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_StreetName.R
\name{match_StreetName}
\alias{match_StreetName}
\title{Match Street Name}
\usage{
match_StreetName(address, StreetMatch2 = NULL, Numbers = NULL, hash = FALSE)
}
\arguments{
\item{address}{A character vector of addresses.}

\item{StreetMatch2}{The integer vector result of \code{match_StreetType(address, m = 2L)}.
If \code{NULL} the vector is evaluated internally.}

\item{Numbers}{If not \code{NULL}, then a list from \code{extract_flatNumberFirstLast}.}

\item{hash}{Whether to hash the result as an integer vector.}
}
\description{
Match Street Name
}
