% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_StreetType.R
\name{match_StreetType}
\alias{match_StreetType}
\title{Find the street type within an address}
\usage{
match_StreetType(address)
}
\arguments{
\item{address}{A character vector, every string an address.}
}
\value{
A list of two elements. The first element are the indices of
street type in \code{.permitted_street_type_ord()} that is found in the
address. The second element are the corresponding string positions of
the street so identified.
}
\description{
Find the street type within an address
}
\examples{
cds <- .permitted_street_type_ord()
head(cds)
match_StreetType("712 FLINDERS STREET MELBOURNE 3004")
#                 012345678901234
match_StreetType("712 FLINDERS ST MELBOURNE 3004")

}
