% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_word.R
\name{match_word}
\alias{match_word}
\title{Find word within a sentence}
\usage{
match_word(x, tbl)
}
\arguments{
\item{x}{A character vector of uppercase sentences.}

\item{tbl}{A table of words. Long vectors are not permitted.}
}
\value{
An integer vector the same length as \code{x}, where the
\code{i}-th entry
is the integer position of the first word in \code{tbl}
detected in \code{x[i]}. Non-matches return \code{NA}. Words
are strings of uppercase separated by spaces.
}
\description{
Find word within a sentence
}
