% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nany_lowercase.R
\name{nany_lowercase}
\alias{nany_lowercase}
\title{Uppercase character vectors}
\usage{
nany_lowercase(x)
}
\arguments{
\item{x}{A character vector, of ASCII elements.}
}
\value{
\describe{
\item{\code{nany_lowercase}}{\code{FALSE} is any char in \code{x} is a lowercase letter.}
}
}
\description{
Ensures all elements of a character vector are uppercase;
no lowercase characters.
}
\examples{
nany_lowercase("ABC")
nany_lowercase("ABC 123 /--")
nany_lowercase("ABC 123 /-- z")

}
