% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ste_fst.R
\name{read_ste_fst}
\alias{read_ste_fst}
\title{Get internal data}
\usage{
read_ste_fst(
  ste = c("ACT", "NSW", "NT", "OT", "QLD", "SA", "TAS", "VIC", "WA"),
  columns = NULL,
  data_env = getOption("healthyAddress.data_env"),
  rbind = TRUE
)
}
\arguments{
\item{ste}{The abbreviated state name.}

\item{columns}{Character vector of columns to select. If \code{NULL}, all columns are selected.}

\item{data_env}{The environment in which objects are cached. Mainly for
internal use.}

\item{rbind}{Whether or not to bind the list result should multiple states
be requested.}
}
\description{
Get internal data
}
