% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toupper_basic.R
\name{toupper_basic}
\alias{toupper_basic}
\title{Uppercase}
\usage{
toupper_basic(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
The same as \code{toupper(x)} for ASCII entries. For implementation
reasons, strings wider than 32767 characters (bytes) will be ignored.
}
\description{
Uppercase
}
