% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_StreetType.R
\name{match_StreetType}
\alias{match_StreetType}
\title{Find the street type within an address}
\usage{
match_StreetType(address, m = 0L)
}
\arguments{
\item{address}{A character vector, every string an address.}

\item{m}{An integer. Values greater than \code{0} are for internal use and
encode additional data in the return value:

\enumerate{
\item An integer vector \code{y} is returned such that \code{y mod 256} gives
which street code and \code{y / 256} gives the character position in address
at which the street type starts.
}}
}
\value{
Which of \code{\link{.permitted_street_type_ord}} appears
in \code{address}.
}
\description{
Find the street type within an address
}
\examples{
cds <- .permitted_street_type_ord()
head(cds)
match_StreetType("712 FLINDERS STREET MELBOURNE 3004")
match_StreetType("712 FLINDERS STREET MELBOURNE 3004", 1L)

}
