% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_address.R
\name{standardize_address}
\alias{standardize_address}
\alias{standard_address2}
\alias{standard_address3}
\title{Standard address}
\usage{
standardize_address(
  Address,
  AddressLine2 = NULL,
  return.type = c("data.table", "integer"),
  integer_StreetType = FALSE,
  StreetHash = FALSE
)

standard_address2(Address)

standard_address3(Line1, Line2, Postcode = NULL, KeepStreetName = FALSE)
}
\arguments{
\item{Address}{A character vector, either a full address or (if \code{AddressLine2}
is not \code{NULL}) the first line of an Australian address.}

\item{AddressLine2}{Either \code{NULL} (the default) or a character vector,
the same length as \code{Address} giving the second line of the Address.}

\item{return.type}{Either \code{"data.table"} or \code{"integer"}.
\code{"data.table"} implies a table of columns separating the address components.
\code{"integer"} means an integer vector creating a bijection between the
address and the \code{PSMA} internal id.}

\item{integer_StreetType}{Should the street type be returned as an integer
vector?}

\item{StreetHash}{Should \code{STREET_NAME} be returned as an integer hash.}

\item{Line1, Line2, Postcode}{For addresses split by line. \code{Line1} is
assumed to end with the street type. The second line is only used to determine
\code{Postcode}, and then only if it is \code{NULL}, the default.}

\item{KeepStreetName}{Should an additional character vector be included in
the result of the street name?}
}
\value{
A \code{data.table} containing columns indicating the components of the standard address,
including \code{H0} which is the \code{\link{HashStreetName}} of the street name.
}
\description{
Standard address
}
