% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_Postcodes.R
\name{unique_Postcodes}
\alias{unique_Postcodes}
\title{Unique postcodes of}
\usage{
unique_Postcodes(x)
}
\arguments{
\item{x}{An integer vector of postcodes.}
}
\value{
A (sorted) integer vector of the unique, non-NA values in \code{x}.
}
\description{
Unique postcodes of
}
